requirements.txt
requests==2.31.0
pandas==2.1.3
numpy==1.24.3
ta==0.10.2
python-dotenv==1.0.0
scikit-learn==1.3.2
nltk==3.8.1
yfinance==0.2.32
ccxt==4.0.17
textblob==0.17.1
